+++ SimulationEngine/Communication/MessageProvider.cs
@@ -7,7 +7,7 @@ namespace SimulationEngine.Communication
         private static int _sequenceNumber;
  
         public static Message CreateMessage(TypeMessage type, string sender, string addressee,
-            string code, IDictionary<string, object> dataParameters, int timestamp)
+            string code, IDictionary<string, object> dataParameters, long timestamp)
         {
             var msg = new Message(type, sender, addressee, code, dataParameters, timestamp)
             {
